pyproject.toml
[tool.poetry]
name = "scythe-cli"
version = "0.15.0"
description = "A Harvest is always better with a good tool"
authors = ["Sean Collings <seanrcollings@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/seanrcollings/scythe"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.25.1"
xdg = "^5.1.1"
pydantic = "^1.9.0"
arc-cli = "8.1.5"
oyaml = "^1.0"
diskcache = "^5.4.0"
textual = {extras = ["dev"], version = "^0.24.0"}
httpx = "^0.24.0"
toml = "^0.10.2"


[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.800"
pylint = "^2.7.0"
types-requests = "^2.27.7"
types-PyYAML = "^6.0.3"


[tool.poetry.scripts]
scythe = "scythe_cli.application:application"

[tool.poetry.group.dev.dependencies]
types-toml = "^0.10.8.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
